update the testing structure for dbt_service
