+++ ddpui/tests/services/dbt_service/test_delete_dbt_workspace.py
@@ -1,0 +1,58 @@
+"""Tests for delete_dbt_workspace function."""
+
+import pytest
+from pathlib import Path
+from unittest.mock import patch, Mock
+from ddpui.ddpdbt.dbt_service import delete_dbt_workspace
+from ddpui.models.org import Org, OrgDbt, OrgPrefectBlockv1
+from ddpui.ddpprefect import DBTCLIPROFILE, SECRET
+
+
+@pytest.mark.django_db
+def test_delete_dbt_workspace():
+    """tests the delete_dbt_workspace function"""
+    org = Org.objects.create(name="temp", slug="temp")
+
+    cli_block = OrgPrefectBlockv1.objects.create(
+        org=org,
+        block_type=DBTCLIPROFILE,
+        block_id="dbtcli-block-id",
+        block_name="dbtcli-block-id",
+    )
+
+    org.dbt = OrgDbt.objects.create(
+        gitrepo_url="gitrepo_url",
+        project_dir="project-dir",
+        target_type="tgt",
+        default_schema="default_schema",
+        cli_profile_block=cli_block,
+    )
+    org.save()
+
+    assert OrgDbt.objects.filter(gitrepo_url="gitrepo_url").count() == 1
+    OrgPrefectBlockv1.objects.create(
+        org=org,
+        block_type=SECRET,
+        block_id="secret-block-id",
+        block_name="secret-git-pull",
+    )
+
+    assert OrgPrefectBlockv1.objects.filter(block_id="dbtcli-block-id").count() == 1
+    assert OrgPrefectBlockv1.objects.filter(block_id="secret-block-id").count() == 1
+
+    with patch("ddpui.ddpdbt.dbt_service.os.path.exists") as mock_exists, patch(
+        "ddpui.ddpdbt.dbt_service.shutil.rmtree"
+    ) as mock_rmtree, patch(
+        "ddpui.ddpdbt.dbt_service.prefect_service.delete_dbt_cli_profile_block"
+    ) as mock_delete_dbt_cli_block, patch(
+        "ddpui.ddpdbt.dbt_service.prefect_service.delete_secret_block"
+    ) as mock_delete_secret_block:
+        delete_dbt_workspace(org)
+        mock_exists.return_value = True
+        mock_rmtree.assert_called_once_with("project-dir")
+        mock_delete_dbt_cli_block.assert_called_once_with("dbtcli-block-id")
+        mock_delete_secret_block.assert_called_once_with("secret-block-id")
+
+    assert org.dbt is None
+    assert OrgDbt.objects.filter(gitrepo_url="gitrepo_url").count() == 0
+    assert OrgPrefectBlockv1.objects.filter(block_id="block-id").count() == 0
